+++ CHANGES.md
@@ -15,6 +15,8 @@
     - @szktty @enm10k
 - [UPDATE] WebRTC 91.4472.9.1 に上げる
     - @enm10k
+- [ADD] AV1 に対応する
+    - @enm10k
 - [ADD] libwebrtc のログレベルを設定する API を追加
     - `Sora.setWebRTCLogLevel(_:)`
     - @szktty
@@ -32,6 +34,12 @@
         - SignalingPush.data
     - 修正にともない、 SignalingClientMetadata が SignalingNotifyMetadata にリネームされました
     - @enm10k
+- [CHANGES] type: notify のシグナリング・メッセージに対応する struct として SignalingNotify を追加する
+    - event_type 毎に定義されていた以下の struct が廃止され、 SignalingNotify に統合されます
+        - SignalingNotifyConnection
+        - SignalingNotifySpotlightChanged
+        - SignalingNotifyNetworkStatus
+    - @enm10k
 - [CHANGE] サイマルキャストのオプションを整理する
     - SimulcastQuality を削除し、 SimulcastRid を追加する
     - Configuration.simulcastQuality を削除し、 simulcastRid を追加する
@@ -45,6 +53,7 @@
     - SignalingNotifyConnection.data
     - @enm10k
 - [FIX] サイマルキャストのパラメーター active: false が無効化されてしまう問題を修正する
+    - @enm10k
 
 ## 2020.7.2