+++ CHANGES.md
@@ -12,9 +12,16 @@
 ## develop
 
 - [CHANGE] 接続開始時のカメラ・デバイスを指定可能にする
-    - Configuration.cameraSettings.position に .front または .back を設定して、接続開始時のカメラ・デバイスを指定します
-    - この変更に伴い、 CameraVideoCapturer の API が破壊的に変更されました
-    - 併せて、 `Configuration.videoCapturerDevice` が廃止されました
+    - `Configuration.cameraSettings.position` に `.front` または `.back` を設定して、接続開始時のカメラ・デバイスを指定します
+    - この修正に伴い、以下の API が変更されました
+        - `CameraVideoCapturer` の API を破壊的に変更
+        - `CameraVideoCapturer.Settings` を `CameraSettings` にリネーム
+        - `VideoCapturerHandlers` を `CameraVideoCapturerHandlers` にリネーム
+        - `VideoCapturer` を廃止
+        - `VideoCapturerDevice` を廃止
+        - `CameraPosition` を廃止
+        - `Configuration.videoCapturerDevice` を廃止
+        - `MediaStream.videoCapturer` を廃止
     - @szktty @enm10k
 
 ## 2021.1