+++ CHANGES.md
@@ -18,6 +18,8 @@
     - @szktty @enm10k
 - [UPDATE] シグナリング・メッセージ re-offer, re-answer に対応する
     - @enm10k
+- [UPDATE] AES-GCM を有効にする
+    - @enm10k
 - [CHANGE] 接続開始時のカメラ・デバイスを指定可能にする
     - `Configuration.cameraSettings.position` に `.front` または `.back` を設定して、接続開始時のカメラ・デバイスを指定します
     - この修正に伴い、以下の API が変更されました