+++ CHANGES.md
@@ -11,6 +11,8 @@
 
 ## develop
 
+- [UPDATE] Swift Package Manager に対応する
+    - @miosakuma
 - [UPDATE] システム条件を変更する
     - iOS 12.1 以降
     - @miosakuma